+++ src/stitching/mergeSchemas.ts
@@ -3,47 +3,37 @@ import {
   FieldNode,
   FragmentDefinitionNode,
   FragmentSpreadNode,
-  GraphQLArgument,
-  GraphQLArgumentConfig,
-  GraphQLCompositeType,
   GraphQLField,
-  GraphQLFieldConfig,
-  GraphQLFieldConfigArgumentMap,
-  GraphQLFieldConfigMap,
   GraphQLFieldMap,
   GraphQLFieldResolver,
-  GraphQLInputField,
-  GraphQLInputFieldConfig,
-  GraphQLInputFieldConfigMap,
-  GraphQLInputFieldMap,
   GraphQLInputObjectType,
   GraphQLInputType,
   GraphQLInterfaceType,
   GraphQLList,
+  GraphQLNamedType,
   GraphQLNonNull,
   GraphQLObjectType,
   GraphQLResolveInfo,
   GraphQLScalarType,
   GraphQLSchema,
   GraphQLType,
-  GraphQLNamedType,
   GraphQLUnionType,
   InlineFragmentNode,
   Kind,
   OperationDefinitionNode,
   SelectionSetNode,
-  TypeNode,
   TypeNameMetaFieldDef,
+  TypeNode,
   VariableDefinitionNode,
   VariableNode,
   buildASTSchema,
   execute,
+  extendSchema,
   getNamedType,
   isCompositeType,
   isNamedType,
   parse,
   visit,
-  extendSchema,
 } from 'graphql';
 import TypeRegistry from './TypeRegistry';
 import { IResolvers } from '../Interfaces';
@@ -52,8 +42,11 @@ import {
   extractExtensionDefinitions,
   addResolveFunctionsToSchema,
 } from '../schemaGenerator';
-import resolveFromParentTypename from './resolveFromParentTypename';
-import defaultMergedResolver from './defaultMergedResolver';
+import {
+  recreateCompositeType,
+  fieldMapToFieldConfigMap,
+} from './schemaRecreation';
+import typeFromAST from './typeFromAST';
 import { checkResultAndHandleErrors } from './errors';
 
 export type MergeInfo = {
@@ -83,7 +76,7 @@ export default function mergeSchemas({
     leftType: GraphQLNamedType,
     rightType: GraphQLNamedType,
   ) => GraphQLNamedType;
-  resolvers: (mergeInfo: MergeInfo) => IResolvers;
+  resolvers?: (mergeInfo: MergeInfo) => IResolvers;
 }): GraphQLSchema {
   if (!onTypeConflict) {
     onTypeConflict = defaultOnTypeConflict;
@@ -96,6 +89,7 @@ export default function mergeSchemas({
   const mergeInfo: MergeInfo = createMergeInfo(typeRegistry);
 
   const actualSchemas: Array<GraphQLSchema> = [];
+  const typeFragments: Array<DocumentNode> = [];
   const extensions: Array<DocumentNode> = [];
   let fullResolvers: IResolvers = {};
 
@@ -108,7 +102,7 @@ export default function mergeSchemas({
         const actualSchema = buildASTSchema(parsedSchemaDocument);
         actualSchemas.push(actualSchema);
       } catch (e) {
-        // Could not create a schema from parsed string, will use extensions
+        typeFragments.push(parsedSchemaDocument);
       }
       parsedSchemaDocument = extractExtensionDefinitions(parsedSchemaDocument);
       if (parsedSchemaDocument.definitions.length > 0) {
@@ -176,6 +170,15 @@ export default function mergeSchemas({
     }
   });
 
+  typeFragments.forEach(document => {
+    document.definitions.forEach(def => {
+      const type = typeFromAST(typeRegistry, def);
+      if (type) {
+        typeRegistry.addType(type.name, type, onTypeConflict);
+      }
+    });
+  });
+
   let passedResolvers = {};
   if (resolvers) {
     passedResolvers = resolvers(mergeInfo);
@@ -225,123 +228,6 @@ export default function mergeSchemas({
   addResolveFunctionsToSchema(mergedSchema, fullResolvers);
 
   return mergedSchema;
-}
-
-function recreateCompositeType(
-  schema: GraphQLSchema,
-  type: GraphQLCompositeType | GraphQLInputObjectType,
-  registry: TypeRegistry,
-): GraphQLCompositeType | GraphQLInputObjectType {
-  if (type instanceof GraphQLObjectType) {
-    const fields = type.getFields();
-    const interfaces = type.getInterfaces();
-
-    return new GraphQLObjectType({
-      name: type.name,
-      description: type.description,
-      isTypeOf: type.isTypeOf,
-      fields: () => fieldMapToFieldConfigMap(fields, registry),
-      interfaces: () => interfaces.map(iface => registry.resolveType(iface)),
-    });
-  } else if (type instanceof GraphQLInterfaceType) {
-    const fields = type.getFields();
-
-    return new GraphQLInterfaceType({
-      name: type.name,
-      description: type.description,
-      fields: () => fieldMapToFieldConfigMap(fields, registry),
-      resolveType: (parent, context, info) =>
-        resolveFromParentTypename(parent, info.schema),
-    });
-  } else if (type instanceof GraphQLUnionType) {
-    return new GraphQLUnionType({
-      name: type.name,
-      description: type.description,
-      types: () =>
-        type.getTypes().map(unionMember => registry.resolveType(unionMember)),
-      resolveType: (parent, context, info) =>
-        resolveFromParentTypename(parent, info.schema),
-    });
-  } else if (type instanceof GraphQLInputObjectType) {
-    return new GraphQLInputObjectType({
-      name: type.name,
-      description: type.description,
-      fields: () => inputFieldMapToFieldConfigMap(type.getFields(), registry),
-    });
-  } else {
-    throw new Error(`Invalid type ${type}`);
-  }
-}
-
-function fieldMapToFieldConfigMap(
-  fields: GraphQLFieldMap<any, any>,
-  registry: TypeRegistry,
-): GraphQLFieldConfigMap<any, any> {
-  const result: GraphQLFieldConfigMap<any, any> = {};
-  Object.keys(fields).forEach(name => {
-    result[name] = fieldToFieldConfig(fields[name], registry);
-  });
-  return result;
-}
-
-function fieldToFieldConfig(
-  field: GraphQLField<any, any>,
-  registry: TypeRegistry,
-): GraphQLFieldConfig<any, any> {
-  return {
-    type: registry.resolveType(field.type),
-    args: argsToFieldConfigArgumentMap(field.args, registry),
-    resolve: defaultMergedResolver,
-    description: field.description,
-    deprecationReason: field.deprecationReason,
-  };
-}
-
-function argsToFieldConfigArgumentMap(
-  args: Array<GraphQLArgument>,
-  registry: TypeRegistry,
-): GraphQLFieldConfigArgumentMap {
-  const result: GraphQLFieldConfigArgumentMap = {};
-  args.forEach(arg => {
-    const [name, def] = argumentToArgumentConfig(arg, registry);
-    result[name] = def;
-  });
-  return result;
-}
-
-function argumentToArgumentConfig(
-  argument: GraphQLArgument,
-  registry: TypeRegistry,
-): [string, GraphQLArgumentConfig] {
-  return [
-    argument.name,
-    {
-      type: registry.resolveType(argument.type),
-      defaultValue: argument.defaultValue,
-      description: argument.description,
-    },
-  ];
-}
-
-function inputFieldMapToFieldConfigMap(
-  fields: GraphQLInputFieldMap,
-  registry: TypeRegistry,
-): GraphQLInputFieldConfigMap {
-  const result: GraphQLInputFieldConfigMap = {};
-  Object.keys(fields).forEach(name => {
-    result[name] = inputFieldToFieldConfig(fields[name], registry);
-  });
-  return result;
-}
-
-function inputFieldToFieldConfig(
-  field: GraphQLInputField,
-  registry: TypeRegistry,
-): GraphQLInputFieldConfig {
-  return {
-    type: registry.resolveType(field.type),
-    description: field.description,
-  };
 }
 
 function createMergeInfo(typeRegistry: TypeRegistry): MergeInfo {
@@ -602,13 +488,13 @@ function filterSelectionSetDeep(
   fragments: Array<FragmentDefinitionNode>;
   usedVariables: Array<string>;
 } {
-  const validFragments: Array<string> = [];
+  const validFragments: Array<FragmentDefinitionNode> = [];
   Object.keys(fragments).forEach(fragmentName => {
     const fragment = fragments[fragmentName];
     const typeName = fragment.typeCondition.name.value;
     const innerType = schema.getType(typeName);
     if (innerType) {
-      validFragments.push(fragment.name.value);
+      validFragments.push(fragment);
     }
   });
   let {
@@ -680,7 +566,7 @@ function filterSelectionSet(
   },
   type: GraphQLType,
   selectionSet: SelectionSetNode,
-  validFragments: Array<string>,
+  validFragments: Array<FragmentDefinitionNode>,
 ): {
   selectionSet: SelectionSetNode;
   usedFragments: Array<string>;
@@ -692,15 +578,9 @@ function filterSelectionSet(
   const filteredSelectionSet = visit(selectionSet, {
     [Kind.FIELD]: {
       enter(node: FieldNode): null | undefined | FieldNode {
-        let parentType: GraphQLType = resolveType(
+        let parentType: GraphQLNamedType = resolveType(
           typeStack[typeStack.length - 1],
         );
-        if (
-          parentType instanceof GraphQLNonNull ||
-          parentType instanceof GraphQLList
-        ) {
-          parentType = parentType.ofType;
-        }
         if (
           parentType instanceof GraphQLObjectType ||
           parentType instanceof GraphQLInterfaceType
@@ -762,8 +642,22 @@ function filterSelectionSet(
       }
     },
     [Kind.FRAGMENT_SPREAD](node: FragmentSpreadNode): null | undefined {
-      if (validFragments.indexOf(node.name.value) !== -1) {
+      const fragmentFiltered = validFragments.filter(
+        frg => frg.name.value === node.name.value,
+      );
+      const fragment = fragmentFiltered[0];
+      if (fragment) {
+        if (fragment.typeCondition) {
+          const innerType = schema.getType(fragment.typeCondition.name.value);
+          const parentType: GraphQLNamedType = resolveType(
+            typeStack[typeStack.length - 1],
+          );
+          if (!implementsAbstractType(parentType, innerType)) {
+            return null;
+          }
+        }
         usedFragments.push(node.name.value);
+        return;
       } else {
         return null;
       }
@@ -772,7 +666,10 @@ function filterSelectionSet(
       enter(node: InlineFragmentNode): null | undefined {
         if (node.typeCondition) {
           const innerType = schema.getType(node.typeCondition.name.value);
-          if (innerType) {
+          const parentType: GraphQLNamedType = resolveType(
+            typeStack[typeStack.length - 1],
+          );
+          if (implementsAbstractType(parentType, innerType)) {
             typeStack.push(innerType);
           } else {
             return null;
@@ -811,6 +708,30 @@ function resolveType(type: GraphQLType): GraphQLNamedType {
     lastType = lastType.ofType;
   }
   return lastType;
+}
+
+function implementsAbstractType(
+  parent: GraphQLType,
+  child: GraphQLType,
+  bail: boolean = false,
+): boolean {
+  if (parent === child) {
+    return true;
+  } else if (
+    parent instanceof GraphQLInterfaceType &&
+    child instanceof GraphQLObjectType
+  ) {
+    return child.getInterfaces().indexOf(parent) !== -1;
+  } else if (
+    parent instanceof GraphQLUnionType &&
+    child instanceof GraphQLObjectType
+  ) {
+    return parent.getTypes().indexOf(child) !== -1;
+  } else if (parent instanceof GraphQLObjectType && !bail) {
+    return implementsAbstractType(child, parent, true);
+  }
+
+  return false;
 }
 
 function typeToAst(type: GraphQLInputType): TypeNode {